factcheck_sidekick/requirements.txt
# Add your project dependencies here
# Example:
langchain==0.3.13
langchain-community==0.3.13
langchain-openai==0.2.14
tavily-python==0.5.0
requests==2.32.3
beautifulsoup4==4.12.3
chromadb==0.6.3
tiktoken==0.8.0
python-dotenv==1.0.1
duckduckgo-search==2.6.1
openai==1.109.1
flask==3.0.0